requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
Pillow
